refactor connection and move mosaic endpoints to extensions
