+++ titiler/pgstac/factory.py
@@ -1,20 +1,9 @@
 """Custom MosaicTiler Factory for PgSTAC Mosaic Backend."""
+
 import os
-import re
 import warnings
 from dataclasses import dataclass
-from typing import (
-    Any,
-    Callable,
-    Dict,
-    Generator,
-    List,
-    Literal,
-    Optional,
-    Tuple,
-    Type,
-    Union,
-)
+from typing import Any, Callable, Dict, List, Literal, Optional, Tuple, Type, Union
 from urllib.parse import urlencode
 
 import jinja2
@@ -24,7 +13,6 @@ from cogeo_mosaic.errors import MosaicNotFoundError
 from fastapi import Body, Depends, HTTPException, Path, Query
 from fastapi.dependencies.utils import get_dependant, request_params_to_args
 from geojson_pydantic import Feature, FeatureCollection
-from psycopg import sql
 from psycopg.rows import class_row
 from pydantic import conint
 from rio_tiler.constants import MAX_THREADS, WGS84_CRS
@@ -32,7 +20,6 @@ from rio_tiler.mosaic.methods.base import MosaicMethodBase
 from starlette.datastructures import QueryParams
 from starlette.requests import Request
 from starlette.responses import HTMLResponse, Response
-from starlette.routing import NoMatchFound
 from starlette.templating import Jinja2Templates
 from typing_extensions import Annotated
 
@@ -56,11 +43,11 @@ from titiler.core.utils import render_image
 from titiler.mosaic.factory import PixelSelectionParams
 from titiler.pgstac import model
 from titiler.pgstac.dependencies import (
-    BackendParams,
     PathParams,
     PgSTACParams,
     SearchParams,
     TileParams,
+    db_conn,
 )
 from titiler.pgstac.mosaic import PGSTACBackend
 
@@ -112,12 +99,12 @@ class MosaicTilerFactory(BaseTilerFactory):
 
     pixel_selection_dependency: Callable[..., MosaicMethodBase] = PixelSelectionParams
 
-    backend_dependency: Type[DefaultDependency] = BackendParams
+    pgstac_dependency: Type[DefaultDependency] = PgSTACParams
+    backend_dependency: Type[DefaultDependency] = DefaultDependency
 
     # Add/Remove some endpoints
     add_statistics: bool = False
     add_viewer: bool = False
-    add_mosaic_list: bool = False
     add_part: bool = False
 
     templates: Jinja2Templates = DEFAULT_TEMPLATES
@@ -145,10 +132,6 @@ class MosaicTilerFactory(BaseTilerFactory):
 
     def register_routes(self) -> None:
         """This Method register routes to the router."""
-        self._search_routes()
-        if self.add_mosaic_list:
-            self._search_list_routes()
-
         # NOTE: `assets` route HAVE TO be registered before `tiles` routes
         self._assets_routes()
 
@@ -217,10 +200,11 @@ class MosaicTilerFactory(BaseTilerFactory):
             color_formula=Depends(ColorFormulaParams),
             colormap=Depends(self.colormap_dependency),
             render_params=Depends(self.render_dependency),
-            pgstac_params: PgSTACParams = Depends(),
+            pgstac_params=Depends(self.pgstac_dependency),
             backend_params=Depends(self.backend_dependency),
             reader_params=Depends(self.reader_dependency),
             env=Depends(self.environment_dependency),
+            connection=Depends(db_conn),
         ):
             """Create map tile."""
             scale = scale or 1
@@ -229,6 +213,7 @@ class MosaicTilerFactory(BaseTilerFactory):
             with rasterio.Env(**env):
                 with self.reader(
                     searchid,
+                    connection=connection,
                     tms=tms,
                     reader_options={**reader_params},
                     **backend_params,
@@ -329,20 +314,20 @@ class MosaicTilerFactory(BaseTilerFactory):
             color_formula=Depends(ColorFormulaParams),
             colormap=Depends(self.colormap_dependency),
             render_params=Depends(self.render_dependency),
-            pgstac_params: PgSTACParams = Depends(),
+            pgstac_params=Depends(self.pgstac_dependency),
             backend_params=Depends(self.backend_dependency),
             reader_params=Depends(self.reader_dependency),
+            connection=Depends(db_conn),
         ):
             """Return TileJSON document for a SearchId."""
-            with request.app.state.dbpool.connection() as conn:
-                with conn.cursor(row_factory=class_row(model.Search)) as cursor:
-                    cursor.execute(
-                        "SELECT * FROM searches WHERE hash=%s;",
-                        (searchid,),
-                    )
-                    search_info = cursor.fetchone()
-                    if not search_info:
-                        raise MosaicNotFoundError(f"SearchId `{searchid}` not found")
+            with connection.cursor(row_factory=class_row(model.Search)) as cursor:
+                cursor.execute(
+                    "SELECT * FROM searches WHERE hash=%s;",
+                    (searchid,),
+                )
+                search_info = cursor.fetchone()
+                if not search_info:
+                    raise MosaicNotFoundError(f"SearchId `{searchid}` not found")
 
             route_params = {
                 "searchid": search_info.id,
@@ -431,7 +416,7 @@ class MosaicTilerFactory(BaseTilerFactory):
             color_formula=Depends(ColorFormulaParams),
             colormap=Depends(self.colormap_dependency),
             render_params=Depends(self.render_dependency),
-            pgstac_params: PgSTACParams = Depends(),
+            pgstac_params=Depends(self.pgstac_dependency),
             backend_params=Depends(self.backend_dependency),
             reader_params=Depends(self.reader_dependency),
             env=Depends(self.environment_dependency),
@@ -488,17 +473,17 @@ class MosaicTilerFactory(BaseTilerFactory):
                 Optional[int],
                 Query(description="Overwrite default maxzoom."),
             ] = None,
+            connection=Depends(db_conn),
         ):
             """OGC WMTS endpoint."""
-            with request.app.state.dbpool.connection() as conn:
-                with conn.cursor(row_factory=class_row(model.Search)) as cursor:
-                    cursor.execute(
-                        "SELECT * FROM searches WHERE hash=%s;",
-                        (searchid,),
-                    )
-                    search_info = cursor.fetchone()
-                    if not search_info:
-                        raise MosaicNotFoundError(f"SearchId `{searchid}` not found")
+            with connection.cursor(row_factory=class_row(model.Search)) as cursor:
+                cursor.execute(
+                    "SELECT * FROM searches WHERE hash=%s;",
+                    (searchid,),
+                )
+                search_info = cursor.fetchone()
+                if not search_info:
+                    raise MosaicNotFoundError(f"SearchId `{searchid}` not found")
 
             route_params = {
                 "searchid": searchid,
@@ -523,7 +508,7 @@ class MosaicTilerFactory(BaseTilerFactory):
                 self.rescale_dependency,
                 self.colormap_dependency,
                 self.render_dependency,
-                PgSTACParams,
+                self.pgstac_dependency,
                 self.reader_dependency,
                 self.backend_dependency,
             ]
@@ -639,14 +624,16 @@ class MosaicTilerFactory(BaseTilerFactory):
                 Literal[tuple(self.supported_tms.list())],
                 f"Identifier selecting one of the TileMatrixSetId supported (default: '{self.default_tms}')",
             ] = self.default_tms,
-            pgstac_params: PgSTACParams = Depends(),
+            pgstac_params=Depends(self.pgstac_dependency),
             backend_params=Depends(self.backend_dependency),
             reader_params=Depends(self.reader_dependency),
+            connection=Depends(db_conn),
         ):
             """Return a list of assets which overlap a given tile"""
             tms = self.supported_tms.get(tileMatrixSetId)
             with self.reader(
                 searchid,
+                connection=connection,
                 tms=tms,
                 reader_options={**reader_params},
                 **backend_params,
@@ -663,13 +650,15 @@ class MosaicTilerFactory(BaseTilerFactory):
             lat: Annotated[float, Path(description="Latitude")],
             searchid=Depends(self.path_dependency),
             coord_crs=Depends(CoordCRSParams),
-            pgstac_params: PgSTACParams = Depends(),
+            pgstac_params=Depends(self.pgstac_dependency),
             backend_params=Depends(self.backend_dependency),
             reader_params=Depends(self.reader_dependency),
+            connection=Depends(db_conn),
         ):
             """Return a list of assets for a given point."""
             with self.reader(
                 searchid,
+                connection=connection,
                 reader_options={**reader_params},
                 **backend_params,
             ) as src_dst:
@@ -679,354 +668,6 @@ class MosaicTilerFactory(BaseTilerFactory):
                     coord_crs=coord_crs or WGS84_CRS,
                     **pgstac_params,
                 )
-
-    def _search_routes(self) -> None:  # noqa: C901
-        """register search routes."""
-
-        @self.router.post(
-            "/register",
-            responses={200: {"description": "Register a Search."}},
-            response_model=model.RegisterResponse,
-            response_model_exclude_none=True,
-        )
-        def register_search(
-            request: Request, search_query=Depends(self.search_dependency)
-        ):
-            """Register a Search query."""
-            search, metadata = search_query
-
-            with request.app.state.dbpool.connection() as conn:
-                with conn.cursor(row_factory=class_row(model.Search)) as cursor:
-                    cursor.execute(
-                        "SELECT * FROM search_query(%s, _metadata => %s);",
-                        (
-                            search.model_dump_json(by_alias=True, exclude_none=True),
-                            metadata.model_dump_json(exclude_none=True),
-                        ),
-                    )
-                    search_info = cursor.fetchone()
-
-            links: List[model.Link] = [
-                model.Link(
-                    rel="metadata",
-                    title="Mosaic metadata",
-                    href=self.url_for(request, "info_search", searchid=search_info.id),
-                ),
-                model.Link(
-                    rel="tilejson",
-                    title="Link for TileJSON",
-                    href=self.url_for(request, "tilejson", searchid=search_info.id),
-                ),
-            ]
-
-            try:
-                links.append(
-                    model.Link(
-                        rel="map",
-                        title="Link for Map viewer",
-                        href=self.url_for(
-                            request, "map_viewer", searchid=search_info.id
-                        ),
-                    )
-                )
-            except NoMatchFound:
-                pass
-
-            try:
-                links.append(
-                    model.Link(
-                        rel="wmts",
-                        title="Link for WMTS",
-                        href=self.url_for(request, "wmts", searchid=search_info.id),
-                    )
-                )
-            except NoMatchFound:
-                pass
-
-            if search_info.metadata.defaults:
-                # List of dependencies a `/tile` URL should validate
-                # Note: Those dependencies should only require Query() inputs
-                tile_dependencies = [
-                    self.layer_dependency,
-                    self.dataset_dependency,
-                    self.pixel_selection_dependency,
-                    self.process_dependency,
-                    self.rescale_dependency,
-                    self.colormap_dependency,
-                    self.render_dependency,
-                    PgSTACParams,
-                    self.reader_dependency,
-                    self.backend_dependency,
-                ]
-
-                for name, values in search_info.metadata.defaults.items():
-                    query_string = urlencode(values, doseq=True)
-                    try:
-                        self.check_query_params(
-                            dependencies=tile_dependencies,
-                            query_params=QueryParams(query_string),
-                        )
-                    except Exception as e:
-                        warnings.warn(
-                            f"Cannot construct URL for layer `{name}`: {repr(e)}",
-                            UserWarning,
-                            stacklevel=2,
-                        )
-                        continue
-
-                    links.append(
-                        model.Link(
-                            title=f"TileJSON link for `{name}` layer.",
-                            rel="tilejson",
-                            href=self.url_for(
-                                request,
-                                "tilejson",
-                                searchid=search_info.id,
-                            )
-                            + f"?{query_string}",
-                        )
-                    )
-
-            return model.RegisterResponse(searchid=search_info.id, links=links)
-
-        @self.router.get(
-            "/{searchid}/info",
-            responses={200: {"description": "Get Search query metadata."}},
-            response_model=model.Info,
-            response_model_exclude_none=True,
-        )
-        def info_search(request: Request, searchid=Depends(self.path_dependency)):
-            """Get Search query metadata."""
-            with request.app.state.dbpool.connection() as conn:
-                with conn.cursor(row_factory=class_row(model.Search)) as cursor:
-                    cursor.execute(
-                        "SELECT * FROM searches WHERE hash=%s;",
-                        (searchid,),
-                    )
-                    search_info = cursor.fetchone()
-
-            if not search_info:
-                raise MosaicNotFoundError(f"SearchId `{searchid}` not found")
-
-            links: List[model.Link] = [
-                model.Link(
-                    rel="self",
-                    title="Mosaic metadata",
-                    href=self.url_for(request, "info_search", searchid=search_info.id),
-                ),
-                model.Link(
-                    title="Link for TileJSON",
-                    rel="tilejson",
-                    href=self.url_for(request, "tilejson", searchid=search_info.id),
-                ),
-            ]
-
-            try:
-                links.append(
-                    model.Link(
-                        rel="map",
-                        title="Link for Map viewer",
-                        href=self.url_for(
-                            request, "map_viewer", searchid=search_info.id
-                        ),
-                    )
-                )
-            except NoMatchFound:
-                pass
-
-            try:
-                links.append(
-                    model.Link(
-                        rel="wmts",
-                        title="Link for WMTS",
-                        href=self.url_for(request, "wmts", searchid=search_info.id),
-                    )
-                )
-            except NoMatchFound:
-                pass
-
-            if search_info.metadata.defaults:
-                for name, values in search_info.metadata.defaults.items():
-                    links.append(
-                        model.Link(
-                            title=f"TileJSON link for `{name}` layer.",
-                            rel="tilejson",
-                            href=self.url_for(
-                                request,
-                                "tilejson",
-                                searchid=search_info.id,
-                            )
-                            + f"?{urlencode(values, doseq=True)}",
-                        )
-                    )
-
-            return model.Info(search=search_info, links=links)
-
-    def _search_list_routes(self) -> None:
-        """Add mosaic listing route."""
-
-        @self.router.get(
-            "/list",
-            responses={200: {"description": "List Mosaics in PgSTAC."}},
-            response_model=model.Infos,
-            response_model_exclude_none=True,
-        )
-        def list_mosaic(
-            request: Request,
-            limit: Annotated[
-                int,
-                Query(
-                    ge=1,
-                    le=1000,
-                    description="Page size limit",
-                ),
-            ] = 10,
-            offset: Annotated[
-                int,
-                Query(
-                    ge=0,
-                    description="Page offset",
-                ),
-            ] = 0,
-            sortby: Annotated[
-                Optional[str],
-                Query(
-                    description="Sort the response items by a property (ascending (default) or descending).",
-                ),
-            ] = None,
-        ):
-            """List a Search query."""
-            # Default filter to only return `metadata->type == 'mosaic'`
-            mosaic_filter = sql.SQL("metadata->>'type' = 'mosaic'")
-
-            # additional metadata property filter passed in query-parameters
-            # <propname>=val - filter for a metadata property. Multiple property filters are ANDed together.
-            qs_key_to_remove = ["limit", "offset", "sortby"]
-            additional_filter = [
-                sql.SQL("metadata->>{key} = {value}").format(
-                    key=sql.Literal(key), value=sql.Literal(value)
-                )
-                for (key, value) in request.query_params.items()
-                if key.lower() not in qs_key_to_remove
-            ]
-            filters = [
-                sql.SQL("WHERE"),
-                sql.SQL("AND ").join([mosaic_filter, *additional_filter]),
-            ]
-
-            def parse_sort_by(sortby: str) -> Generator[sql.Composable, None, None]:
-                """Parse SortBy expression."""
-                for s in sortby.split(","):
-                    parts = re.match(
-                        "^(?P<dir>[+-]?)(?P<prop>.*)$", s.lstrip()
-                    ).groupdict()  # type:ignore
-
-                    prop = parts["prop"]
-                    if parts["prop"] in ["lastused", "usecount"]:
-                        prop = sql.Identifier(prop)
-                    else:
-                        prop = sql.SQL("metadata->>{}").format(sql.Literal(prop))
-
-                    if parts["dir"] == "-":
-                        order = sql.SQL("{} DESC").format(prop)
-                    else:
-                        order = sql.SQL("{} ASC").format(prop)
-
-                    yield order
-
-            # sortby=[+|-]PROP - sort the response items by a property (ascending (default) or descending).
-            order_by = []
-            if sortby:
-                sort_expr = list(parse_sort_by(sortby))
-                if sort_expr:
-                    order_by = [
-                        sql.SQL("ORDER BY"),
-                        sql.SQL(", ").join(sort_expr),
-                    ]
-
-            with request.app.state.dbpool.connection() as conn:
-                with conn.cursor() as cursor:
-                    # Get Total Number of searches rows
-                    query = [
-                        sql.SQL("SELECT count(*) FROM searches"),
-                        *filters,
-                    ]
-                    cursor.execute(sql.SQL(" ").join(query))
-                    nb_items = int(cursor.fetchone()[0])
-
-                    # Get rows
-                    cursor.row_factory = class_row(model.Search)
-                    query = [
-                        sql.SQL("SELECT * FROM searches"),
-                        *filters,
-                        *order_by,
-                        sql.SQL("LIMIT %(limit)s OFFSET %(offset)s"),
-                    ]
-                    cursor.execute(
-                        sql.SQL(" ").join(query), {"limit": limit, "offset": offset}
-                    )
-                    searches_info = cursor.fetchall()
-
-            qs = QueryParams({**request.query_params, "limit": limit, "offset": offset})
-            links = [
-                model.Link(
-                    rel="self",
-                    href=self.url_for(request, "list_mosaic") + f"?{qs}",
-                ),
-            ]
-
-            if len(searches_info) < nb_items:
-                next_token = offset + len(searches_info)
-                qs = QueryParams(
-                    {**request.query_params, "limit": limit, "offset": next_token}
-                )
-                links.append(
-                    model.Link(
-                        rel="next",
-                        href=self.url_for(request, "list_mosaic") + f"?{qs}",
-                    ),
-                )
-
-            if offset > 0:
-                prev_token = offset - limit if (offset - limit) > 0 else 0
-                qs = QueryParams(
-                    {**request.query_params, "limit": limit, "offset": prev_token}
-                )
-                links.append(
-                    model.Link(
-                        rel="prev",
-                        href=self.url_for(request, "list_mosaic") + f"?{qs}",
-                    ),
-                )
-
-            return model.Infos(
-                searches=[
-                    model.Info(
-                        search=search,
-                        links=[
-                            model.Link(
-                                rel="metadata",
-                                href=self.url_for(
-                                    request, "info_search", searchid=search.id
-                                ),
-                            ),
-                            model.Link(
-                                rel="tilejson",
-                                href=self.url_for(
-                                    request, "tilejson", searchid=search.id
-                                ),
-                            ),
-                        ],
-                    )
-                    for search in searches_info
-                ],
-                links=links,
-                context=model.Context(
-                    returned=len(searches_info),
-                    matched=nb_items,
-                    limit=limit,
-                ),
-            )
 
     def _statistics_routes(self):
         """Register /statistics endpoint."""
@@ -1057,10 +698,11 @@ class MosaicTilerFactory(BaseTilerFactory):
             pixel_selection=Depends(self.pixel_selection_dependency),
             stats_params=Depends(self.stats_dependency),
             histogram_params=Depends(self.histogram_dependency),
-            pgstac_params: PgSTACParams = Depends(),
+            pgstac_params=Depends(self.pgstac_dependency),
             backend_params=Depends(self.backend_dependency),
             reader_params=Depends(self.reader_dependency),
             env=Depends(self.environment_dependency),
+            connection=Depends(db_conn),
         ):
             """Get Statistics from a geojson feature or featureCollection."""
             fc = geojson
@@ -1070,6 +712,7 @@ class MosaicTilerFactory(BaseTilerFactory):
             with rasterio.Env(**env):
                 with self.reader(
                     searchid,
+                    connection=connection,
                     reader_options={**reader_params},
                     **backend_params,
                 ) as src_dst:
@@ -1137,15 +780,17 @@ class MosaicTilerFactory(BaseTilerFactory):
             color_formula=Depends(ColorFormulaParams),
             colormap=Depends(self.colormap_dependency),
             render_params=Depends(self.render_dependency),
-            pgstac_params: PgSTACParams = Depends(),
+            pgstac_params=Depends(self.pgstac_dependency),
             backend_params=Depends(self.backend_dependency),
             reader_params=Depends(self.reader_dependency),
             env=Depends(self.environment_dependency),
+            connection=Depends(db_conn),
         ):
             """Create image from a bbox."""
             with rasterio.Env(**env):
                 with self.reader(
                     searchid,
+                    connection=connection,
                     reader_options={**reader_params},
                     **backend_params,
                 ) as src_dst:
@@ -1212,15 +857,17 @@ class MosaicTilerFactory(BaseTilerFactory):
             color_formula=Depends(ColorFormulaParams),
             colormap=Depends(self.colormap_dependency),
             render_params=Depends(self.render_dependency),
-            pgstac_params: PgSTACParams = Depends(),
+            pgstac_params=Depends(self.pgstac_dependency),
             backend_params=Depends(self.backend_dependency),
             reader_params=Depends(self.reader_dependency),
             env=Depends(self.environment_dependency),
+            connection=Depends(db_conn),
         ):
             """Create image from a geojson feature."""
             with rasterio.Env(**env):
                 with self.reader(
                     searchid,
+                    connection=connection,
                     reader_options={**reader_params},
                     **backend_params,
                 ) as src_dst:
