+++ titiler/pgstac/main.py
@@ -25,6 +25,11 @@ from titiler.mosaic.errors import MOSAIC_STATUS_CODES
 from titiler.pgstac import __version__ as titiler_pgstac_version
 from titiler.pgstac.db import close_db_connection, connect_to_db
 from titiler.pgstac.dependencies import ItemPathParams
+from titiler.pgstac.extensions import (
+    searchInfoExtension,
+    searchListExtension,
+    searchRegisterExtension,
+)
 from titiler.pgstac.factory import MosaicTilerFactory
 from titiler.pgstac.reader import PgSTACReader
 from titiler.pgstac.settings import ApiSettings, PostgresSettings
@@ -103,8 +108,12 @@ mosaic = MosaicTilerFactory(
     router_prefix="/mosaic",
     add_statistics=True,
     add_viewer=True,
-    add_mosaic_list=True,
     add_part=True,
+    extensions=[
+        searchRegisterExtension(),
+        searchInfoExtension(),
+        searchListExtension(),
+    ],
 )
 app.include_router(mosaic.router, tags=["Mosaic"], prefix="/mosaic")
